requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
boltons==21.0.0
bracex==2.6
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
cryptography==46.0.4
cssselect2==0.9.0
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.135.3
fonttools==4.62.0
glom==25.12.0
googleapis-common-protos==1.72.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.4
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mcp==1.23.3
mdurl==0.1.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==26.0
peewee==3.19.0
pillow==12.1.1
protobuf==6.33.5
pycparser==3.0
pydantic[email]==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
PyMySQL==1.1.2
pyphen==0.17.2
python-dotenv==1.2.1
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pywin32==311; sys_platform == "win32"
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rich==13.5.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
semantic-version==2.10.0
semgrep==1.151.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.2.0
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.1.0
tomli==2.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcmatch==8.5.2
weasyprint==68.1
webencodings==0.5.1
wrapt==1.17.3
zipp==3.23.0
zopfli==0.4.1
